Refactor Asset.cs into individual Asset/ class files
This makes classes a lot easier to navigate in the VS solution explorer.
