+++ DropGrid.Client/Asset/AssetBundle.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DropGrid.Client.Asset
+{
+    /// <summary>
+    /// A collection of assets to be used for a specific portion of game logic.
+    /// Generally we want to load as little redundant assets as possible.
+    /// </summary>
+    public class AssetBundle
+    {
+
+        public void add(Asset asset)
+        {
+
+        }
+    }
+}
