+++ DropGrid.Client/Asset/AssetLoader.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DropGrid.Client.Asset
+{
+    class AssetLoader
+    {
+    }
+}
